add JWT token

+++ routes/api.php
@@ -2,6 +2,8 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\UsersController;
+use App\Http\Controllers\AuthController;
 
 /*
 |--------------------------------------------------------------------------
@@ -14,15 +16,36 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-//Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
-//    return $request->user();
-//});
+Route::prefix('users')->group(function () {
+    //Prefijo V1, todo lo que este dentro de este grupo se accedera escribiendo v1 en el navegador, es decir /api/v1/*
+    Route::post('login', [AuthController::class, 'authenticate']);
+    Route::group(['middleware' => ['auth:api']], function() {
+        //Todo lo que este dentro de este grupo requiere verificación de usuario.
+        Route::post('logout', [AuthController::class, 'logout']);
+        Route::post('create', [UsersController::class, 'create'])->middleware(['admin']);
+        Route::get('read/{id?}', [UsersController::class, 'read']);
+        Route::post('update', [UsersController::class, 'update'])->middleware(['admin']);
+        Route::delete('delete/{id}', [UsersController::class, 'delete'])->middleware(['admin']);
+    });
+});
+/*
 Route::group([
+    'middleware' => (['auth:api']),
     'prefix' => 'users',
     'controller' => App\Http\Controllers\UsersController::class,
+    
 ], function () {
-    Route::post('create', 'create');
+    Route::post('create', 'create')->middleware(['admin']);
+    Route::post('update', 'update')->middleware(['admin']);
     Route::get('read/{id?}', 'read');
-    Route::post('update', 'update');
-    Route::delete('delete/{id}', 'delete');
+    Route::delete('delete/{id}', 'delete')->middleware(['admin']);
+});   
+
+Route::group([
+    'prefix' => 'users',
+    'controller' => App\Http\Controllers\AuthController::class
+    
+], function () {
+    Route::post('login', 'authenticate');
 });
+*/
